@@ -1,8 +1,9 @@
 {
-  "name": "ivanc-web",
+  "name": "bbytes-web",
   "version": "1.0.0",
   "description": "Personal website built with React",
   "scripts": {
+    "dev": "webpack --progress --display-error-details --config webpack.dev.config.js --watch",
     "heroku-postbuild": "npm run build",
     "build": "webpack --config ./webpack.prod.config.js --progress --colors",
     "start": "node server.js"
@@ -11,22 +12,47 @@
   "license": "Apache 2.0",
   "devDependencies": {
     "react-hot-loader": "^1.3.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.10.0"
+    "redux-devtools": "3.3.1",
+    "redux-devtools-dock-monitor": "1.1.1",
+    "redux-devtools-log-monitor": "1.1.1",
+    "redux-logger": "2.7.4",
+    "redux-mock-store": "1.2.1",
+    "webpack-dev-middleware": "1.9.0",
+    "webpack-hot-middleware": "2.13.2"
   },
   "dependencies": {
+    "autoprefixer": "6.5.4",
     "babel-core": "^6.4.5",
     "babel-loader": "^6.2.1",
-    "babel-preset-react": "^6.3.13",
+    "babel-preset-es2015": "6.18.0",
+    "babel-preset-es2015-loose": "8.0.0",
+    "babel-preset-react": "6.16.0",
+    "babel-preset-stage-1": "6.16.0",
     "css-loader": "^0.23.1",
-    "es6-promise": "^3.1.2",
+    "clean-webpack-plugin": "0.1.14",
+    "es6-promise": "4.0.5",
     "exports-loader": "^0.6.3",
     "express": "^4.13.4",
+    "extract-text-webpack-plugin": "2.1.0",
+    "file-loader": "0.9.0",
+    "html-webpack-plugin": "2.24.1",
     "imports-loader": "^0.6.5",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "isomorphic-fetch": "2.2.1",
+    "node-sass": "3.13.1",
+    "postcss-import": "9.0.0",
+    "postcss-loader": "1.2.0",
+    "react": "15.4.2",
+    "react-addons-css-transition-group": "^0.14.1",
+    "react-dom": "15.4.2",
+    "react-redux": "4.4.6",
+    "react-router": "3.0.0",
+    "react-router-redux": "4.0.7",
+    "redux": "3.6.0",
+    "redux-thunk": "2.1.0",
+    "sass-loader": "4.1.0",
     "style-loader": "^0.13.1",
-    "webpack": "^1.12.11",
+    "tcomb-form": "0.9.10",
+    "webpack": "2.2.0-rc.1",
     "whatwg-fetch": "^0.11.0"
   }
 }
