@@ -1,13 +1,33 @@
-var React = require('react');
-var ReactDOM = require('react-dom');
-var WebApp = require('./WebApp.js');
-var GAnalytics = require('./GAnalytics.js');
-//Include styles using webpack
-require("../styles/main.css");
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { browserHistory } from 'react-router';
+import { syncHistoryWithStore } from 'react-router-redux';
 
-ReactDOM.render(
-  <WebApp />,
-  document.getElementById('root')
+import Root from './containers/Root/Root';
+import configureStore from './store/configureStore';
+import { authLoginUserSuccess } from './actions/auth';
+
+
+const initialState = {};
+const target = document.getElementById('root');
+
+const store = configureStore(initialState, browserHistory);
+const history = syncHistoryWithStore(browserHistory, store);
+
+const node = (
+    <Root store={store} history={history}/>
 );
 
-GAnalytics.trackPageView();
+const token = sessionStorage.getItem('token');
+let user = {};
+try {
+    user = JSON.parse(sessionStorage.getItem('user'));
+} catch (e) {
+    // Failed to parse
+}
+
+if (token !== null) {
+    store.dispatch(authLoginUserSuccess(token, user));
+}
+
+ReactDOM.render(node, target);
